requirements.txt
Django>=4.2,<5.1
wagtail>=6.1,<6.2
drf_spectacular==0.27.2
drf_yasg
Pillow==10.3.0
setuptools
virtualenv==20.24.6
flake8==6.1.0
isort==5.12.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pyproject-api==1.6.1
pytest==7.4.3
python-decouple==3.8
tox==4.11.3
uWSGI==2.0.23
djangorestframework-simplejwt==5.2.2
django-cors-headers
djangorestframework